+++ docs/index.rst
@@ -11,7 +11,7 @@ recognition methods geared torwards the analysis of biosignals.
 
 Highlights:
 
--  Support for various biosignals: BVP, ECG, EDA, EEG, EMG, Respiration
+-  Support for various biosignals: PPG, ECG, EDA, EEG, EMG, Respiration
 -  Signal analysis primitives: filtering, frequency analysis
 -  Clustering
 -  Biometrics
